feat(backyourstack-dispatcher): refactor and rename
